requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0

langgraph>=0.1.0
langchain>=0.2.0
langchain-google-genai>=1.0.0
google-generativeai>=0.5.0

chromadb>=0.5.0

confluent-kafka>=2.4.0

pydantic>=2.7.0
pydantic-settings>=2.2.0

ragas>=0.1.0

streamlit>=1.35.0

grpcio>=1.63.0
grpcio-tools>=1.63.0

httpx>=0.27.0
python-dotenv>=1.0.0

pytest>=8.2.0
pytest-asyncio>=0.23.0
